perf(cuda): fp32 expf in MoE routing — eager pin owns byte identity (post-bf93510)
Reverts k_moe_route's correctly-rounded (float)exp((double)x) to plain
device expf, keeping the reciprocal-multiply mirror. Rationale: since
bf93510 the certification harnesses pin RUNNER_MOE_EAGER=1, so the MoE
byte-identity property is defined over the eager host-routing path; the
double-exp's sole purpose — bit-matching correctly-rounded host libms —
is void on the fast-math (libmvec) cert box and buys nothing certified.

Property downgrade, stated plainly: the fused default is no longer
byte-identical to the host routing even on UCRT-class hosts. Its
contract is the weaker verified class — expert selection identical,
selw within 1 ulp of the host reference given the same logits —
re-verified on the 3070 via RUNNER_DEBUG_MOE after the revert.
docs/compatibility-program.md amended accordingly (its "byte-identical
on correctly-rounded hosts" sentence described the double-exp build).

Gates (3070): make test green incl. drift; certified (eager-pinned)
CPU==GPU identity byte-green on Qwen3-30B and gemma-4-26B over 128
greedy tokens; bench md5 unchanged (6bd488bd).

+++ src/kernels.cu
@@ -2215,18 +2215,19 @@ extern "C" __global__ void k_moe_route(const float *logits, int *sel,
         if (lg[e] > mx) mx = lg[e];
     float ssum = 0.0f;
     for (int e = 0; e < ne; e++) {
-        // (float)exp((double)x), NOT device expf: device expf differs from
-        // the host libm by ~1-2 ulp, and that difference alone broke the
-        // certified full-offload CPU==GPU byte identity (near-tie flip;
-        // isolated 2026-07-29 — selection was identical, selw differed in
-        // the last bit). Rounding the double exp to float yields the
-        // correctly-rounded float exp in all but ~2^-28 of calls (CUDA
-        // documents double exp at 1 ulp), which bit-matches a
-        // correctly-rounded host expf (verified exhaustively-sampled
-        // against UCRT; glibc scalar expf is also correctly rounded). The
-        // serial one-thread contract above already makes this kernel
-        // latency-bound, so the fp64 cost is noise.
-        float p = (float)exp((double)(lg[e] - mx));
+        // Device expf differs from the host libm by ~1-2 ulp, so the fused
+        // path is NOT byte-identical to the host routing — the certified
+        // MoE byte-identity property is therefore defined over the EAGER
+        // path (RUNNER_MOE_EAGER=1, pinned in the certification harnesses;
+        // see docs/compatibility-program.md). The fused default's contract
+        // is the weaker, verified class: expert selection identical, selw
+        // within 1 ulp of the host reference (the reciprocal-multiply
+        // mirror below keeps the divergence to expf alone). History: a
+        // correctly-rounded (float)exp((double)x) here DID bit-match
+        // correctly-rounded hosts (UCRT), but a fast-math/libmvec host is
+        // ~4 ulp and unreachable by construction, so certification pinned
+        // eager and this kernel keeps the plain fp32 expf.
+        float p = expf(lg[e] - mx);
         lg[e] = p;
         ssum += p;
     }
